`NULL` pointer checking for `malloc` and `realloc` and index formatting for relevant functions

+++ c_dynamically_sized_arrays/src/dynamic_array.h
@@ -16,7 +16,7 @@ typedef struct dynamic_array_t {
 
 dynamic_array_t* make_dynamic_arr(size_t length);
 size_t dynamic_arr_at(dynamic_array_t* array, int index);
-void dynamic_arr_set(dynamic_array_t* array, size_t index, size_t value);
+void dynamic_arr_set(dynamic_array_t* array, int index, size_t value);
 void dynamic_arr_resize(dynamic_array_t* array, size_t length);
 void dynamic_arr_append(dynamic_array_t* array, size_t value);
 void dynamic_arr_pop(dynamic_array_t* array);
